src/server/consumer/requirements.txt
confluent-kafka==2.4.0
psycopg2==2.9.9
PyYAML==6.0.1
